Remoção e criação de playlists adaptadas a relação ouvinte-playlist

+++ autofalante/views.py
@@ -6,6 +6,7 @@ from django.shortcuts import render
 from django.utils import timezone
 from django.http import HttpResponse
 from django.template import loader
+from django.core.exceptions import ValidationError
 from django.contrib.auth.models import User
 from .models import Playlist, Musica, Listener
 from .forms import PlaylistCreationForm, PlaylistDeletionForm
@@ -31,7 +32,7 @@ def home(request):
     songs = Musica.objects.filter(playlist = ALL_SONGS_PLAYLIST_ID)
     if (request.method == 'POST' and 'create' in request.POST):
         creation_form = PlaylistCreationForm(request.POST)
-        deletion_form = PlaylistDeletionForm()
+        deletion_form = PlaylistDeletionForm(request = request)
         if (creation_form.is_valid()):
             new_playlist = Playlist(
                 playlist_title = creation_form.cleaned_data['playlist_name'],
@@ -41,15 +42,15 @@ def home(request):
             u.listener.playlists.add(new_playlist)
             creation_form = PlaylistCreationForm()
     elif(request.method == 'POST' and 'delete' in request.POST):    
-        deletion_form = PlaylistDeletionForm(request.POST)
+        deletion_form = PlaylistDeletionForm(data = request.POST, request = request)
         creation_form = PlaylistCreationForm()
-        if (deletion_form.is_valid()):
+        if (deletion_form.is_valid() and owned_playlists.filter(playlist_title = deletion_form.cleaned_data['playlist_name']).count()!= 0):
             deleted_playlist = Playlist.objects.get(playlist_title = deletion_form.cleaned_data['playlist_name'])
             deleted_playlist.delete()
-            deletion_form = PlaylistDeletionForm()
+            deletion_form = PlaylistDeletionForm(request = request)
     else:
         creation_form = PlaylistCreationForm()
-        deletion_form = PlaylistDeletionForm()            
+        deletion_form = PlaylistDeletionForm(request = request)            
     template = loader.get_template('autofalante/home.html')
     context = {
         'owned_playlists': owned_playlists,
